python/requirements.txt
Flask==1.1.1
waitress
mariadb
mysql-connector-python